requirements.txt
jamo>=0.4.1
inflect>=7.0.0
unidecode>=1.3.6
pyopenjtalk>=0.3.1
jieba>=0.42.1
pypinyin>=0.49.0
termcolor>=1.1.0
numpy>=1.23.5
scikit-learn>=1.1.1
torchinfo>=1.8.0
tqdm>=4.65.0
visualdl>=2.5.3
torchaudio>=0.13.0
librosa>=0.10.1
matplotlib==3.3.1
soundfile>=0.12.1
pyyaml>=5.4.1
monotonic-align
cn2an
eng_to_ipa
ko_pron
phonemizer
indic_transliteration
num_thai
opencc
